+++ Tree.hpp
@@ -26,23 +26,23 @@ struct hash_vector {
 
 class Node{
 public:
-  double _x;
-  double _y;
+  float _x;
+  float _y;
   Node* _parent;
-  vector<double> _vec;
-  Node(double x, double y, Node* parent);
+  vector<float> _vec;
+  Node(float x, float y, Node* parent);
   //Definir operação de soma
   Node operator + (const Node& obj);
   //Definir operação de subtração
   Node operator - (const Node& obj);
   //Definir operação de produto escalar
-  double operator * (const Node& obj);
+  float operator * (const Node& obj);
   //comparar dois nodes
   bool operator == (const Node& obj);
   //Módulo do vetor;
-  double modulus() const;
+  float modulus() const;
   //Multiplicar por constante
-  Node multiplyByConstant(double c) const;
+  Node multiplyByConstant(float c) const;
   //Calcular unitário
   Node makeUnitary();
   //destrutor
@@ -51,22 +51,22 @@ public:
 
 class Tree{
 private:
-    double _probGoal;
-    double _probWaypoint;
+    float _probGoal;
+    float _probWaypoint;
     Node _root;
     Node _goal;
     Environment& _env;
     ObstacleGrid& _obs;
-    flann::Index<flann::L2_Simple<double>> _kdtree;
+    flann::Index<flann::L2_Simple<float>> _kdtree;
     vector<Node> _waypointCache;
 
 public:
-  unordered_map<vector<double>, Node, hash_vector> _nodemap;
+  unordered_map<vector<float>, Node, hash_vector> _nodemap;
 
-  Tree(Node& root, Node& goal, double probGoal, double probWaypoint, Environment& env, ObstacleGrid& obs, vector<Node>& waypointCache);
+  Tree(Node& root, Node& goal, float probGoal, float probWaypoint, Environment& env, ObstacleGrid& obs, vector<Node>& waypointCache);
   //função para determinar a extensão da árvore
   //ponteiro para determinar  distancia do ponto recem adiciondo ao objetivo
-  bool extend(double step, Node* last = nullptr);
+  bool extend(float step, Node* last = nullptr);
   //função para escolher target
   Node chooseTarget();
   //calcular Neighrest Neighbor
@@ -76,7 +76,7 @@ public:
   //encontrar em ordem o caminho que leva ao objetivo
   vector<Node> backtrack();
   // funcao que faz crecer a arvore proceduralmente
-  bool grow(double step = 8, double threshold = 8);
+  bool grow(float step = 8, float threshold = 8);
   //destrutor
   ~Tree();
 };
